test: Move testcases for autodoc events to test_ext_autodoc_events
